Week7/serverless/serverless_microservices/orders/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.28.62
pydantic==2.4.2
mangum==0.17.0
python-dotenv==1.0.0
